Updated createGroupOnOgranisation to follow same pattern as other groups creation; renamed Resolvers to Queries which is what they are
