+++ src/schema/resolvers/index.ts
@@ -1,4 +1,3 @@
 export * from './CreateMutations';
 export * from './UpdateMutations';
-export * from './Resolvers';
-export * from './OrganisationMutations';
+export * from './Queries';
