+++ src/services/index.ts
@@ -1,0 +1,3 @@
+export * from './OrganisationService';
+export * from './EcoverseService';
+export * from './ChallengeService';
